feat: refactor audio, notes, and theme handling; add UI for instruments and drones; migrate to ES6 modules.
