+++ index.html
@@ -131,6 +131,6 @@
   <footer>
     <p>&copy; 2025 Your Name</p>
   </footer>
-  <script src="assets/script.js"></script>
+  <script type="module" src="assets/js/main.js"></script>
 </body>
 </html>
